csv_uploader_api/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
pandas
python-multipart
passlib[bcrypt]
